# swagger.yaml
openapi: 3.0.0
info:
  title: Petit Creux
  version: 1.0.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
tags:
  - name: User
  - name: Recipes
paths:
  /api/v1/auth/register:
    post:
      tags:
        - User
      summary: register
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: demo
                email: demo@gmail.com
                password: password
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/auth/login:
    post:
      tags:
        - User
      summary: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: demo@gmail.com
                password: password
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/auth/:
    get:
      tags:
        - User
      summary: get user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - User
      summary: edit
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: luffy
                email: luffy@gmail.com
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - User
      summary: delete
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/recipes/public:
    post:
      tags:
        - Recipes
      summary: create
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                title: Chocolate Cake
                description: A delicious chocolate cake recipe.
                ingredients:
                  - name: Flour
                    quantity: 200
                    unit: grammes
                  - name: Sugar
                    quantity: 100
                    unit: grammes
                  - name: Butter
                    quantity: 150
                    unit: grammes
                imageUrl: >-
                  https://imgs.search.brave.com/_7JzXdeDFAH-C6S3l6zCnG8o4GO1MUOtcLK8y-_JtvY/rs:fit:500:0:0:0/g:ce/aHR0cHM6Ly9zdG9y/YWdlLmdvb2dsZWFw/aXMuY29tL21lZGlh/LW5ld3Npbml0aWF0/aXZlL2ltYWdlcy9H/TzgwMV9HTklfVmVy/aWZ5aW5nUGhvdG9z/X0NhcmQxX0ltYWdl/MS5vcmlnaW5hbC5q/cGc
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/recipes/userrecipes:
    get:
      tags:
        - Recipes
      summary: get recipe
      security:
        - bearerAuth: []
      parameters:
        - name: createdBy
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
  /api/v1/recipes/:
    delete:
      tags:
        - Recipes
      summary: delete
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}